rename packages sfc-end to sfc-jump
