+++ src/main/packages/components.ts
@@ -64,7 +64,7 @@ import { BPMNPoolComponent } from './bpmn/bpmn-pool/bpmn-pool-component';
 import { BPMNSwimlaneComponent } from './bpmn/bpmn-swimlane/bpmn-swimlane-component';
 import { BPMNDataStoreComponent } from './bpmn/bpmn-data-store/bpmn-data-store-component';
 import { SfcTransitionComponent } from './sfc/sfc-transition/sfc-transition-component';
-import { SfcJumpComponent } from './sfc/sfc-end/sfc-jump-component';
+import { SfcJumpComponent } from './sfc/sfc-jump/sfc-jump-component';
 import { SfcStartComponent } from './sfc/sfc-start/sfc-start-component';
 import { SfcStepComponent } from './sfc/sfc-step/sfc-step-component';
 import { SfcActionTableComponent } from './sfc/sfc-action-table/sfc-action-table-component';
